Feature/chapter 11 supabase

+++ chapter-8/src/app/admin/posts/[id]/page.tsx
@@ -4,10 +4,11 @@ import PostForm from "@/app/_components/PostForm";
 import { usePost } from "@/app/_hooks/usePost";
 import { useParams, useRouter } from "next/navigation";
 import { NextApiPost } from "@/app/_types/typePost";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
 type PostData = {
   title: string;
   content: string;
-  thumbnailUrl: string;
+  thumbnailImageKey: string;
   categories: {
     id: number;
     name: string;
@@ -19,18 +20,27 @@ export default function Home() {
   const id = Number(params.id);
   const { post } = usePost(id);
   const router = useRouter();
+  const { token } = useSupabaseSession();
   const handleSubmit = async (data: PostData) => {
+    if (!token) return;
     await fetch(`/api/admin/posts/${id}`, {
       method: "PUT",
       body: JSON.stringify(data),
+      headers: {
+        Authorization: token,
+      },
     });
     router.push("/admin/posts");
     router.refresh();
   };
   const handleDelete = async () => {
+    if (!token) return;
     if (!confirm("削除しますか？")) return;
     await fetch(`/api/admin/posts/${id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: token,
+      },
     });
     router.push("/admin/posts");
     router.refresh();
@@ -39,7 +49,7 @@ export default function Home() {
     return {
       title: post.title,
       content: post.content,
-      thumbnailUrl: post.thumbnailUrl,
+      thumbnailImageKey: post.thumbnailImageKey,
       // postCategories(入れ子) を categories(フラットな配列) に変換
       categories: post.postCategories.map((pc) => ({
         id: pc.category.id,
